pyproject.toml
[project]
name = "curated-breast-cancer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "gdown>=5.2.0",
    "matplotlib>=3.9.2",
    "onnxruntime>=1.19.2",
    "opencv-python-headless>=4.10.0.84",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "tensorboard>=2.15.2",
    "tensorflow-datasets>=4.9.6",
    "tensorflow==2.15.0",
    "tf2onnx>=1.8.4",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
